\docType{data}
\name{exp1_specificity}
\alias{exp1_specificity}
\title{exp1_specificity dataset}
\format{An object of class \code{\link{list}} related to peptides
  quantification. It consists of qPLEX-RIME data of 10 samples 
  divided into two conditions (ER and IgG).}
\usage{
data(exp1_specificity)
}
\value{
An object of class \code{\link{list}} related to peptides
quantification.
}
\description{
In this experiment we have used the qPLEX-RIME approach to identify ER specific interactors. 
We performed replicate ER RIME pull-downs in five independent biological replicates and an 
equal number of matched IgG mock samples was included.
}
\keyword{data}
\keyword{datasets}

