\docType{data}
\name{exp4_OHT_FP}
\alias{exp4_OHT_FP}
\title{exp4_OHT_FP dataset}
\format{An object of class \code{\link{list}} related to peptides
  quantification. It consists of total proteome data from two experimental
  runs. Each run contains 10 samples divided into four conditions (vehicle,
  tam.2h, tam.6h and tam.24h).}
\usage{
data(exp4_OHT_FP)
}
\value{
An object of class \code{\link{list}} related to peptides
quantification.
}
\description{
We performed two 10plex-TMT time-course experiments to study the 
effect of 4-hydrotamoxifen (OHT) on total protein levels. MCF7 
cells were treated with OHT for 2h, 6h, 24h or for 24h with the 
drug-vehicle alone (ethanol) and a total number of four biological 
replicates were obtained. This is a timecourse experiment 
to study the effect of tamoxifen on full proteome.
}
\keyword{data}
\keyword{datasets}

