\docType{data}
\name{exp5_PDX}
\alias{exp5_PDX}
\title{exp5_PDX dataset}
\format{An object of class \code{\link{list}} related to peptides
  quantification. It consists of qPLEX-RIME data of 10 samples 
  divided into two conditions (PDX and IgG).}
\usage{
data(exp5_PDX)
}
\value{
An object of class \code{\link{list}} related to peptides
quantification.
}
\description{
An ER qPLEX-RIME experiment was performed using three independent 
ER+ human Patient Derived Xenograft (PDX) tumour material. Cryosections 
of each tumour were double-crosslinked and each tumour was split in two 
parts that were used for ER and IgG RIME pull-down assays. One of the 
tumours was split in three different parts to be used as ER or IgG 
qPLEX-RIME in order to assess technical variability.
}
\keyword{data}
\keyword{datasets}

