\docType{data}
\name{exp6_ER}
\alias{exp6_ER}
\title{exp6_ER dataset}
\format{An object of class \code{\link{list}} related to peptides
  quantification. It consists of qPLEX-RIME data of 10 samples 
  divided into two conditions (ER and IgG).}
\usage{
data(exp6_ER)
}
\value{
An object of class \code{\link{list}} related to peptides
quantification.
}
\description{
An ER qPLEX-RIME experiment was performed using five independent 
ER-positive human breast cancer tumours. Cryosections of each tumour 
were double-crosslinked and each tumour was split in two parts that 
were used for ER and IgG RIME pull-down assays.
}
\keyword{data}
\keyword{datasets}

