\docType{data}
\name{exp7_NCOA3}
\alias{exp7_NCOA3}
\title{exp7_NCOA3 dataset}
\format{An object of class \code{\link{list}} related to peptides
  quantification. It consists of qPLEX-RIME data of 10 samples 
  divided into two conditions (NCOA3 and IgG).}
\usage{
data(exp7_NCOA3)
}
\value{
An object of class \code{\link{list}} related to peptides
quantification.
}
\description{
In this experiment we have used the qPLEX-RIME method to identify and
characterize NCOA3 (SRC-3) associated proteins. We performed NCOA3 
RIME pull-downs in five independent biological replicates and in five
matched IgG mock samples.
}
\keyword{data}
\keyword{datasets}

