\docType{data}
\name{exp8_CBP}
\alias{exp8_CBP}
\title{exp8_CBP dataset}
\format{An object of class \code{\link{list}} related to peptides
  quantification. It consists of qPLEX-RIME data of 10 samples 
  divided into two conditions (CBP and IgG).}
\usage{
data(exp8_CBP)
}
\value{
An object of class \code{\link{list}} related to peptides
quantification.
}
\description{
A qPLEX-RIME experiment was designed for the characterization of the CBP 
(CREB-binding protein) interactome. Five independent biological replicates 
of CBP RIME pull-downs and five IgG RIME pull-downs were prepared for this 
experiment.
}
\keyword{data}
\keyword{datasets}

