\docType{data}
\name{exp9_PolII}
\alias{exp9_PolII}
\title{exp9_PolII dataset}
\format{An object of class \code{\link{list}} related to peptides
  quantification. It consists of qPLEX-RIME data of 10 samples 
  divided into two conditions (PolII and IgG).}
\usage{
data(exp9_PolII)
}
\value{
An object of class \code{\link{list}} related to peptides
quantification.
}
\description{
The qPLEX-RIME method was applied for the characterization of the largest 
and catalytic component of RNA polymerase II (RPB1). Particularly, the 
phosphorylated form at Serine 5 in the C-terminal domain (CTD) was used 
as the bait protein. Five biological replicates of RNA polymerase II RIME 
pull-downs and five IgG pull-downs were included for the identification 
and characterization of RNA polymerase II-associated proteins.
}
\keyword{data}
\keyword{datasets}

